backend/requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langsmith
langchain-community
langchain-core
langchain-openai
notebook
trustcall
langgraph-cli[inmem]

flask
flask-cors
dotenv

PyPDF2
pytesseract
pdf2image
Pillow

google-cloud-language